requirements.txt
torch
torchvision
fastapi
uvicorn
gradio
pillow
numpy
scikit-learn
